Move vaultTimeout service into it's own folder
