+++ libs/common/src/services/vaultTimeout/vaultTimeout.service.ts
@@ -1,18 +1,18 @@
-import { AuthService } from "../abstractions/auth.service";
-import { CipherService } from "../abstractions/cipher.service";
-import { CollectionService } from "../abstractions/collection.service";
-import { CryptoService } from "../abstractions/crypto.service";
-import { FolderService } from "../abstractions/folder/folder.service.abstraction";
-import { KeyConnectorService } from "../abstractions/keyConnector.service";
-import { MessagingService } from "../abstractions/messaging.service";
-import { PlatformUtilsService } from "../abstractions/platformUtils.service";
-import { PolicyService } from "../abstractions/policy/policy.service.abstraction";
-import { SearchService } from "../abstractions/search.service";
-import { StateService } from "../abstractions/state.service";
-import { TokenService } from "../abstractions/token.service";
-import { VaultTimeoutService as VaultTimeoutServiceAbstraction } from "../abstractions/vaultTimeout/vaultTimeout.service";
-import { AuthenticationStatus } from "../enums/authenticationStatus";
-import { PolicyType } from "../enums/policyType";
+import { AuthService } from "../../abstractions/auth.service";
+import { CipherService } from "../../abstractions/cipher.service";
+import { CollectionService } from "../../abstractions/collection.service";
+import { CryptoService } from "../../abstractions/crypto.service";
+import { FolderService } from "../../abstractions/folder/folder.service.abstraction";
+import { KeyConnectorService } from "../../abstractions/keyConnector.service";
+import { MessagingService } from "../../abstractions/messaging.service";
+import { PlatformUtilsService } from "../../abstractions/platformUtils.service";
+import { PolicyService } from "../../abstractions/policy/policy.service.abstraction";
+import { SearchService } from "../../abstractions/search.service";
+import { StateService } from "../../abstractions/state.service";
+import { TokenService } from "../../abstractions/token.service";
+import { VaultTimeoutService as VaultTimeoutServiceAbstraction } from "../../abstractions/vaultTimeout/vaultTimeout.service";
+import { AuthenticationStatus } from "../../enums/authenticationStatus";
+import { PolicyType } from "../../enums/policyType";
 
 export class VaultTimeoutService implements VaultTimeoutServiceAbstraction {
   private inited = false;
